Ported Open Simplex to module
