+++ src/lib.rs
@@ -35,7 +35,6 @@ pub use cell::{range_sqr_euclidian2, range_sqr_euclidian3, range_sqr_euclidian4}
 pub use math::{Point2, Point3, Point4};
 
 pub use modules::*;
-pub use open_simplex::{open_simplex2, open_simplex3, open_simplex4};
 pub use perlin::{perlin2, perlin3, perlin4};
 pub use permutationtable::PermutationTable;
 pub use value::{value2, value3, value4};
@@ -47,7 +46,6 @@ mod permutationtable;
 mod brownian;
 mod perlin;
 mod value;
-mod open_simplex;
 mod cell;
 mod modules;
