+++ src/modules/generators/mod.rs
@@ -10,6 +10,7 @@ pub use self::checkerboard::*;
 pub use self::constant::*;
 pub use self::cylinders::*;
 pub use self::fractals::*;
+pub use self::open_simplex::*;
 pub use self::perlin::*;
 pub use self::worley::*;
 
@@ -17,5 +18,6 @@ mod constant;
 mod checkerboard;
 mod cylinders;
 mod fractals;
+mod open_simplex;
 mod perlin;
 mod worley;
